Carlos suggestions (#342)
* style: simplify composition

* style: simplify composition

* style: rename directory

* style: small tweaks on modal
