--- apps/client/src/common/components/input/color-picker-input/PopoverPicker.module.scss
@@ -1,37 +1,0 @@
-@use "../../../../theme/v2Styles" as *;
-
-.picker {
-  width: 300px;
-  background: #f0f0f4;
-  border-radius: 9px;
-}
-
-.reactcolorful {
-  width: auto !important;
-}
-
-.swatch {
-  width: 28px;
-  height: 28px;
-  border-radius: $component-border-radius-md;
-  border: 1px solid #303030;
-  box-shadow: 0 0 0 1px rgba(0, 0, 0, 0.1), inset 0 0 0 1px rgba(0, 0, 0, 0.1);
-  cursor: pointer;
-}
-
-.picker__swatches {
-  display: flex;
-  padding-top: 12px;
-  flex-wrap: wrap;
-}
-
-.picker__swatch {
-  width: 24px;
-  height: 24px;
-  margin: 4px;
-  border: none;
-  padding: 0;
-  border-radius: $component-border-radius-sm;
-  cursor: pointer;
-  outline: none;
-}
