+++ apps/client/src/common/components/input/popover-picker/PopoverPicker.module.scss
@@ -1,0 +1,18 @@
+@use "../../../../theme/_ontimeColours" as *;
+@use "../../../../theme/_v2Styles" as *;
+
+.swatch {
+  width: 32px;
+  height: 32px;
+  border-radius: 16px;
+  border: 1px solid $gray-300;
+  cursor: pointer;
+
+  transition-property: box-shadow;
+  transition-duration: $transition-time-action;
+}
+
+.swatch:hover {
+  box-shadow: 0 0 0 1px $blue-300;
+}
+
